+++ fair/session.py
@@ -76,6 +76,7 @@ class FAIR:
         user_config: str = None,
         debug: bool = False,
         server_mode: fdp_serv.SwitchMode = fdp_serv.SwitchMode.NO_SERVER,
+        server_port: int = 8000,
         testing: bool = False,
     ) -> None:
         """Initialise instance of FAIR sync tool
@@ -97,10 +98,10 @@ class FAIR:
             run in verbose mode
         server_mode : fair.registry.server.SwitchMode, optional
             stop/start server mode during session
+        server_port : int, optional
+            port to run local registry on, default is 8000
         testing : bool
             run in testing mode
-        generate_config : bool
-            if the specified config.yaml does not exist generate it
         """
         if debug:
             logging.getLogger("FAIRDataPipeline").setLevel(logging.DEBUG)
@@ -135,6 +136,7 @@ class FAIR:
         if not os.path.exists(fdp_com.global_config_dir()):
             self._logger.debug("Creating directory: %s", fdp_com.global_config_dir())
             os.makedirs(fdp_com.global_config_dir())
+            assert os.path.exists(fdp_com.global_config_dir())
 
         # Initialise all configuration status dictionaries
         self._local_config: typing.Dict[str, typing.Any] = {}
@@ -156,7 +158,7 @@ class FAIR:
 
         self._load_configurations()
 
-        self._setup_server()
+        self._setup_server(server_port)
 
     def push(self, remote: str = "origin"):
         _staged_data_products = self._stager.get_item_list(True, "data_product")
@@ -170,7 +172,6 @@ class FAIR:
     def purge(
         self,
         verbose: bool = True,
-        local_cfg: bool = False,
         global_cfg: bool = False,
         clear_data: bool = False,
         clear_all: bool = False,
@@ -179,8 +180,6 @@ class FAIR:
 
         Parameters
         ==========
-        local_cfg : bool, optional
-            remove local directories, default is False
         global_cfg : bool, optional
             remove global directories, default is False
         verbose : bool, optional
@@ -198,13 +197,13 @@ class FAIR:
                 click.echo(f"Removing directory '{_root_dir}'")
             shutil.rmtree(_root_dir)
         if clear_all:
-            if verbose:
+            if verbose and os.path.exists(fdp_com.USER_FAIR_DIR):
                 click.echo(f"Removing directory '{fdp_com.USER_FAIR_DIR}'")
             shutil.rmtree(fdp_com.USER_FAIR_DIR)
             return
         if (clear_data or clear_all):
             try:
-                if verbose:
+                if verbose and os.path.exists(fdp_com.default_data_dir()):
                     click.echo(f"Removing directory '{fdp_com.default_data_dir()}'")
                 if os.path.exists(fdp_com.default_data_dir()):
                     shutil.rmtree(fdp_com.default_data_dir())
@@ -220,13 +219,13 @@ class FAIR:
             if os.path.exists(_global_dirs):
                 shutil.rmtree(_global_dirs)
 
-    def _setup_server(self) -> None:
+    def _setup_server(self, port: int) -> None:
         """Start or stop the server if required"""
         self._logger.debug(f"Running server setup for run mode {self._run_mode}")
         if self._run_mode == fdp_serv.SwitchMode.CLI:
-            self._setup_server_cli_mode()
+            self._setup_server_cli_mode(port)
         elif self._run_mode == fdp_serv.SwitchMode.USER_START:
-            self._setup_server_user_start()
+            self._setup_server_user_start(port)
         elif self._run_mode in [
             fdp_serv.SwitchMode.USER_STOP,
             fdp_serv.SwitchMode.FORCE_STOP,
@@ -250,7 +249,7 @@ class FAIR:
             force=self._run_mode == fdp_serv.SwitchMode.FORCE_STOP,
         )
 
-    def _setup_server_cli_mode(self):
+    def _setup_server_cli_mode(self, port: int) -> None:
         self.check_is_repo()
         _cache_addr = os.path.join(
             fdp_com.session_cache_dir(), f"{self._session_id}.run"
@@ -260,7 +259,7 @@ class FAIR:
         # If there are no session cache files start the server
         if not glob.glob(os.path.join(fdp_com.session_cache_dir(), "*.run")):
             self._logger.debug("No sessions found, launching server")
-            fdp_serv.launch_server()
+            fdp_serv.launch_server(port=port)
 
         self._logger.debug(f"Creating new session #{self._session_id}")
 
@@ -273,13 +272,13 @@ class FAIR:
         # Create new session cache file
         pathlib.Path(_cache_addr).touch()
 
-    def _setup_server_user_start(self):
+    def _setup_server_user_start(self, port: int) -> None:
         if not os.path.exists(fdp_com.session_cache_dir()):
             os.makedirs(fdp_com.session_cache_dir())
 
         _cache_addr = os.path.join(fdp_com.session_cache_dir(), "user.run")
 
-        if "registries" not in self._global_config:
+        if self._global_config and "registries" not in self._global_config:
             raise fdp_exc.CLIConfigurationError(
                 "Cannot find server address in current configuration",
                 hint="Is the current location a FAIR repository?",
@@ -289,7 +288,7 @@ class FAIR:
             raise fdp_exc.UnexpectedRegistryServerState("Server already running.")
         click.echo("Starting local registry server")
         pathlib.Path(_cache_addr).touch()
-        fdp_serv.launch_server(fdp_conf.get_local_uri(), verbose=True)
+        fdp_serv.launch_server(port=port, verbose=True)
 
     def run_job(
         self, bash_cmd: str = "", mode: fdp_run.CMD_MODE = fdp_run.CMD_MODE.RUN