+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 - Add `config.yaml` file validation.
+- Add initialisation from existing registry.
+- Switch to setting port not local URI during initialisation.
+- Add option to specify port on `fair registry start`.
 
 # 2021-11-17 [v0.2.3](https://github.com/FAIRDataPipeline/FAIR-CLI/releases/tag/v0.2.3)
 - Move handling of the user `config.yaml` file to a separate class `JobConfiguration`.